temperature_sensor/requirements.txt
blinker==1.9.0
cachetools==5.3.2
certifi==2026.2.25
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.1.3
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
httplib2==0.22.0
idna==3.11
importlib-metadata==6.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pyparsing==3.1.1
python-dotenv==0.21.1
requests==2.33.0
rsa==4.9
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.6.3
w1thermsensor==2.3.0
Werkzeug==3.1.7
zipp==3.23.0

# OpenTelemetry (optional - for distributed tracing)
opentelemetry-api>=1.31.0
opentelemetry-sdk>=1.31.0
opentelemetry-exporter-otlp-proto-grpc>=1.31.0
opentelemetry-instrumentation-flask>=0.52b0
